+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### vNEXT
 
-- Added `ApolloQueryObservable` to support `Rx.Observable` in `Angular2Apollo.watchQuery` method ([PR #51](https://github.com/apollostack/angular2-apollo/pull/51))
+- Added `ApolloQueryObservable` to support `Rx.Observable` in `Angular2Apollo.watchQuery` method ([PR #52](https://github.com/apollostack/angular2-apollo/pull/52))
+- Added `query` method to `Angular2Apollo` service ([PR #51](https://github.com/apollostack/angular2-apollo/pull/51))
 
 ### v0.4.0